Itrombley/update m98

+++ docs/examples/Pixmap_shiftPerPixel.cpp
@@ -1,6 +1,6 @@
 // Copyright 2019 Google LLC.
 // Use of this source code is governed by a BSD-style license that can be found in the LICENSE file.
-#include "fiddle/examples.h"
+#include "tools/fiddle/examples.h"
 // HASH=bf31ee140e2c163c3957276e6d4c4f0c
 REG_FIDDLE(Pixmap_shiftPerPixel, 256, 256, true, 0) {
 const char* color_type(SkColorType ct) {
@@ -29,7 +29,7 @@ void draw(SkCanvas* canvas) {
                                    kGray_8_SkColorType,    kRGBA_F16_SkColorType } ) {
         SkPixmap pixmap(info.makeColorType(colorType), nullptr, 4);
         SkDebugf("color: k" "%s" "_SkColorType" "%*s" "bytesPerPixel: %d shiftPerPixel: %d\n",
-                color_type(colorType), 10 - strlen(color_type(colorType)), " ",
+                color_type(colorType), (int)(10 - strlen(color_type(colorType))), " ",
                 pixmap.info().bytesPerPixel(), pixmap.shiftPerPixel());
     }
 }
